SystemUI: fix recents task view height calculation
Initialize search bar height early enough for layouts to catch it and
not cause jumpiness in certain situations. For instance on a device with
a soft navigation bar, hitting recents from any application, right after
the animation would finish, it would relize its layout was incorrect and
relayout the thumbnails causing a jump.

Change-Id: Id3b5688cabc361db5e874f3f0426c333c006ba2b
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: Vibhor Chaudhary <cvibhu0009@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/RecentsActivity.java